feat: deleteSharedEvent() 에러처리 추가

+++ backend/src/main/java/endolphin/backend/domain/shared_event/SharedEventService.java
@@ -39,6 +39,9 @@ public class SharedEventService {
     }
 
     public void deleteSharedEvent(Long sharedEventId) {
+        if (!sharedEventRepository.existsById(sharedEventId)) {
+            throw new ApiException(ErrorCode.SHARED_EVENT_NOT_FOUND);
+        }
         sharedEventRepository.deleteById(sharedEventId);
     }
